requirements.txt
sqlalchemy
discord.py
mysql-connector
trueskill
python-dotenv
beautifultables
apscheduler
